requirements.txt
paho-mqtt==1.5.0
PyYAML==5.3.1
boto3==1.9.253
